Rename nQueen_AllSolution.java to Nqueen_AllSolution.java
The n-queen puzzle is a problem of placing n queens on a(n×n) chessboard. In such a way, that no two queens can attack each other. Given an integer n, find all the distinct solutions to this "n-queen puzzle". Each solution contains distinct board configurations of the n-queens placement. Where the solution is a permutation of [1,2,3..n] in increasing order. Here the number in the ith place denotes that the ith-column of queen placed in the row with that number if its not posible print -1

Input Format

Given n value is number of queen you want to place

Constraints

1 ≤ n ≤ 20

Output Format

4

Sample Input 0

2
Sample Output 0

-1
Sample Input 1

4
Sample Output 1

- Q - - 
- - - Q 
Q - - - 
- - Q - 

- - Q - 
Q - - - 
- - - Q 
- Q - -